moved get_user_reservation to reservations.py
