+++ lib/database/showtimes.py
@@ -6,16 +6,6 @@ from .connection import connection
 
 
 @gen.coroutine
-def get_reservation_for_user(id):
-    conn = yield connection()
-    result = yield r.table('showtimes').\
-        filter({"user_id": id}).run(conn)
-    if result.items:
-        return result.items[0]
-    return None
-
-
-@gen.coroutine
 def get_showtimes():
     conn = yield connection()
     result = yield r.table('showtimes').order_by('date').run(conn)
